test: cover unified date helpers
